Fix update of binary blobs on Mimer SQL
For long columns, parameter data is supplied to the driver after exec using
SQLPutData. However, some drivers needs to [know at `SQLExecute` time how
much data will be sent through `SQLPutData` for any long columns](https://learn.microsoft.com/en-us/sql/odbc/reference/develop-app/sending-long-data).
Such drivers return 'Y' to a `SQLGetInfo` query for
`SQL_NEED_LONG_DATA_LEN`, although note that SQLServer is one such driver
but it is crying wolf as in practice it works fine if the length is not
provided at exec and it later respects the length passed to `SQLPutData`
without error. Mimer SQL wants to know the length, and really does need it.
Mimer expects that the length buf is set to `SQL_LEN_DATA_AT_EXEC(length)`
at the time of the `SQLExecute` call. If the length buffer is instead set
to `SQL_DATA_AT_EXEC` on exec, this is the sames as `SQL_LEN_DATA_AT_EXEC(0)`,
so it sets the blob to empty, effectively ignoring the length passed to
`SQLPutData` and truncating the data without raising the [expected
22001/'String data, right truncation' error documented for this case](https://learn.microsoft.com/en-us/sql/odbc/reference/syntax/sqlputdata-function).
For most DB's we don't need to provide the length at exec, but the cost
of doing so is very small so it is not worth the extra complexity of
specialising for `SQL_NEED_LONG_DATA_LEN` drivers.