Rewrite all uses of UInt32/Int32 in Webview wrappers to use UInt32Bytes
UInt32/Int32 are types of External.Structure. These have a level of
indirection between the class and buffer to allow the body to be either
a ByteArray buffer, or a pointer to memory. This provides full flexibility
in FFIs where the structure may be in the Smalltalk heap, or in external
memory. However, this is not required in COM wrapper methods that allocate
a buffer and call some COM object method to populate that buffer. In this
case the buffer bytes are always in a Smalltalk managed heap. The wrapper
methods are more efficient if we use an inline buffer, as this both avoids
the needs to dereference the indirection in the FFI call, and requires
only a single memory allocation rather than two.

The AX type library analyser needs to be changed to generate these wrapper
methods in more efficient form, but for now just use the rewrite tool to
convert them.