Remove use of ATL from VM
Most impactful on the in-proc COM stubs. These were broken anyway for
various reasons, so fix them and make the COM random stream sample work
when deployed.

Starting a dolphin dev image inside an app that doesn't include v6 common
controls does not work well. It can be loaded, but a lot of UI does not
work well as it relies on newer common controls. In particular the Task
Dialog does not work, so any message boxes (including the prompt to save
the image when exiting) fail. This needs to be addressed by creating an
activation context for the Dolphin thread in the in-proc stubs.