Remove enhanced instruction set property so builds for SSE2 rather than AVX
AVX has been around since 2011, and has been superceded by a couple of
newer generations since (AVXbut still isn't supported in emulation on
Apple M4 (or earlier). The benefit of AVX in Dolphin is very small, and
mostly limited to a few floating point operations such as the 'floor'
primitive. In practice the performance difference is likely to be
unmeasurable for most, let alone noticeable, so it seems better to err on
the side of running on a wider range of hosts. The option to use AVX
is still available by changing the code generation settings in the build.

The default in VS2022 is to build for SSE2, which is the enhanced
instruction set generation before AVX. Rather than specify this explicitly
we'll just stick with the default, and let Microsoft decide when best to
advance the default as they've done in the past.