Compiler generates better code for initializing objects of known size
The compiler generates a `rep stosd` to initialize pointer objects with
nils. Although faster than it used to be on newer processors, there may
still be better choices for small counts, especially of one! In some cases
the VM knows the fixed size of objects it is allocating, e.g. Character
objects.