Fix #1319: DB Connection buffer classes must be retained on deployment
DBColAttr class>>#initialize is the only reference in code to the
database buffer classes, and so must not be stripped or the buffer classes
will be removed.

A new package closure test is included, but unfortunately the closure test
mechanism isn't sophisticated enough to catch this particular fault
because it only finds missing code caused by the calculation of the
transitive closure of packages required for an application. It does not
account for code removed by the deployment process that iteratively removes
(supposedly) unreferenced methods and classes, as this is difficult to
simulate. It looks possible to create a better test simulation using
RB environments than the simplistic approach currently employed, but it is
not easy, and in the meantime this corrects the particular case of the
apps using Database Connection not working after deployment.