Highlight unbound classes when browsing app contents from deployment log
After deploying an application, one can browser the final content of the
application by opening an Environment Browser that is configured to show
the included code from the manifest that is written to the deployment log.
This would include any unbound classes, but they are not marked out in any
way so they cannot be identified. One has to refer back to the text of the#
log and look for classes in the manifest with the attribute unbound='true'.

With this commit, the unbound classes will be shown in a different colour
and underlined. All apps will have 3 unbound clases in them that are
related to the deployment infrastructure: BasicImageStrippr, ImageStripper,
and DeploymentContext. Any other unbound classes in an application are
likely to cause runtime issues, and so need to be investigated.