Add FileWriteStream for writing to write-only Win32 files
FileWriteStream's raison d'être is to provide buffered serial output
to underlying Win32 "files" that are not readable, e.g. CONOUT$. This
is not supported by Dolphin's standard FileStream as it is designed
to operate in read/write mode regardless of the mode in which the
underlying File is opened, and indeed will fail if open on a File that
is not readable. This is because it attempts to maintain a page buffer
with current file contents in order to support random access (i.e. it
is positionable).

FileWriteStream is a non-readable stream for buffered serial I/O to a
File, and only ever attempts to write to that File. FileWriteStream is
significantly simpler that FileStream, and can be marginally faster
when writing new files. The performance difference is unlikely to be
significant, however, since FileStream is already very efficient.

The Stream hierarchy is quite significantly reorganised, and likewise
the associated tests.