Related to #1319, avoid the need to explicitly preserver DB buffer classes
Rather than initializing the buffer classes into a class variable that is
only accessed from one method, inline the initialization into a constant
expression in that method. The compiler will insert all the classes
referenced in the const expression into the literal frame of the method,
so they will be implicitly preserved, and no special categorisation or
other trick is needed to stop them being removed on application deployment.

Also rename the #bind: selectors so that each operation has a distinct
message name that is more intention revealing.