Fix decoding of last character on page boundary of UTF-8 file with BOM
Bug introduced by a212761dacfe7112975e800c0ca5276f512ad1fa, which changed
the primitiveNext to directly decode to U+FFFDs (replacement chars)
instead of failing on incorrect/incomplete encodings. We need the
primitive to fail on at least incomplete encodings in order that the
primitive can be used with buffers representing a partial range of data
from some underlying store, as otherwise incomplete characters at the
end of the buffer are converted to U+FFFDs rather than triggering
fallback processing that is able to read the encoded character
straddling a page boundary.