UnitTest plugin gets increasingly slow as more windows are opened
The mechanism used to keep the display refreshed when running tests
pumps messages from the input queue, but when a large number of windows are
open this may take a long time to quiesce. Add a new InputState method
to run only the current backlog of messages, not any that might arrive
after starting on the current batch.