Fix #1341: `asOrderedCollection` of UTF string can contain partial chars
Dolphin 8 has a bug that it is converting the string to a sequence of
Characters representing the code units of the UTF-8 encoding. The bug was
introduced by f94156b. The tests accompanying the change don't properly
cover UTF encoding strings. Here we fix the conversion and improve the
tests.