GuiInputState>>topLevelWindows not enumerating child views
An old bug that reoccurred with the high-DPI changes was that this method
ws not enumerating child views on the desktop, e.g. those open for editing
in a ViewComposer. This was caused by the introduction of a a separate
design desktop view, breaking the method's assumption that there is one
unique desktop view instance that could be used to identify top-level views
by identity comparison against the parent. Although this didn't break
preservation of edited child views in VCs across image saves as previously,
that is only because the windows to save are now enumerated with
`toplevelWindowsInZOrder`. The latter uses a Windows API to enumerate the
top-level windows, which is much slower but preserves the z-order.