Fix #1345: Callback thunks not correctly initialized for cdecl convention
Under cdecl calling convention the caller pops arguments from the stack
(to allow for vararg functions). This differs from the normal Windows
calling convention (stdcall) where the callee pops the arguments. Perhaps
surprisingly the bug in the initialization of callback thunks for the cdecl
convention has not been noticed before, but it fails to clear the args size
from previous use of the thunk, so will fail randomly with some kind of
access violation or stack fault.

# Conflicts:
#	Core/Object Arts/Dolphin/Base/ExternalCallback.cls