Remove obsolete background dwell timer from ScintillaView
Originally Scintilla used to run a 100mS interval timer to manage all
features that required some kind of regular background activity, e.g.
hovering events, idle-time syntax colouring, etc. This mechanism prevented
a Dolphin image from quiescing properly, since it would be forced to wake
up regularly to process the WM_TIMER messages. Consequently a workaround
was put in place to switch off the Scintilla timer when the view lost
focus, unless hover events were still wanted even when the view was in the
background. The timer was then restored when the view regained focus.
In 2014, the underlying Scintilla mechanism was improved, see
https://github.com/mirror/scintilla/commit/a4d36c2b76d188199d9b16b4556d9aa88c412731
With this change Scintilla uses separate timers for each feature and turns
them on and off as required. The new timers were based on the
SetCoalescableTimer API introduced in Windows 8. The old 100mS timer was
retained in the code for platforms not supporting the API, although today
any such OS version has long gone out of support so it is effectively
unused. Dolphin continued to switch on the old timer when ScintillaViews
gain focus, generating needless regular timer events. When a very large
number of ScintillaViews is open, the timer events become so frequent in
aggregate (as there is 100mS timer per view), that the system may be
prevented from going idle at all, and so it stops performing some necessary
background processing (such as managing the enablement of toolbar commands)
and this can make the system appear sluggish or unresponsive. This
behaviour does suggest that the background command enablement is flawed in
that it should not cease to work when the system is busy, but correcting
that requires some thought. In the meantime we can just remove the obsolete
timer management from Dolphin.

# Conflicts:
#	Core/Object Arts/Dolphin/IDE/Base/Tools.SmalltalkWorkspace.cls
#	Core/Object Arts/Dolphin/MVP/Views/Scintilla/Dolphin Scintilla View (Deprecated).pax
#	Core/Object Arts/Dolphin/MVP/Views/Scintilla/Dolphin Scintilla View.pax
#	Core/Object Arts/Dolphin/MVP/Views/Scintilla/UI.Scintilla.ScintillaView.cls