CodeMentor plugin info browser picks up dark mode from defaults
The CodeMentorPlugin info browser uses a WebView2 instance with the default
profile. This causes it to pick up whatever the preferred colour scheme
is for the default profile, which is quite likely to be dark. The html
generated by the plugin does not work well in dark mode, and in any case
it doesn't follow the general dolphin IDE theme.

The code mentor may also generated very large amounts of HTML when run
across the entire image for certain rules, e.g. unused methods. This causes
a failure from the WebView2 when trying to get it to display the html text
as it has an odd limitation to 2Mb of html when set from a string. To
workaround this, a limited stream is used to detect exceeding the limit so
that a cut-down info page can be displayed. Displaying a list of 10,000+
method links is not very useful anyway.