Methods sent to self in ##()'s regarded as unimplemented by SmallLint
Sends to self in ##() expressions are sent to the class of the method in
which they are compiled, but SmallLint rule thinks these are self-sends
and so expected them to be implemented on the instance side in instance
methods, which is generally not the case. This then generates fale positive
reports of unimplemented methods.