Maintain same singleton instance of DynamicLinkLibrary classes
Sometimes it may be useful to store a DLL object into an instance variable,
e.g. to make it easy to substitute a mock for testing, but one then needs
to ensure that the variable is re-initialized if persisted, e.g. across
image save and restore. If we maintain the "same" singleton instance
across image saves (or simulate that with a swapping become), then other
refs don't need to be re-initialized.